Permissao e papeis usuarios

+++ app/Http/Controllers/Admin/UsuarioController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Auth;//Sistema de autentificação do Laravel
 use App\User;
+use App\Papel;
 
 class UsuarioController extends Controller
 {
@@ -34,15 +35,25 @@ class UsuarioController extends Controller
     }
     public function index()
     {
-      $usuarios = User::all();
-      return view('admin.usuarios.index', compact('usuarios'));
+      if(auth()->user()->can('usuario_listar')){
+        $usuarios = User::all();
+        return view('admin.usuarios.index', compact('usuarios'));
+      }else{
+        return redirect()->route('admin.principal');
+      }
     }
     public function adicionar()
     {
+      if(!auth()->user()->can('usuario_adicionar')){
+        return redirect()->route('admin.principal');
+      }
       return view('admin.usuarios.adicionar');
     }
     public function salvar(Request $request)
     {
+      if(!auth()->user()->can('usuario_adicionar')){
+        return redirect()->route('admin.principal');
+      }
       $dados = $request->all();
       $usuario = new User();
       $usuario->name = $dados['name'];
@@ -58,11 +69,17 @@ class UsuarioController extends Controller
     }
     public function editar($id)
     {
+      if(!auth()->user()->can('usuario_editar')){
+        return redirect()->route('admin.principal');
+      }
       $usuario = User::find($id);
       return view('admin.usuarios.editar', compact('usuario'));
     }
     public function atualizar(Request $request,$id)
     {
+      if(!auth()->user()->can('usuario_editar')){
+        return redirect()->route('admin.principal');
+      }
       $usuario = User::find($id);
       $dados = $request->all();
       if(isset($dados['password']) && strlen($dados['password']) > 5)
@@ -81,6 +98,9 @@ class UsuarioController extends Controller
 
     public function deletar($id)
     {
+      if(!auth()->user()->can('usuario_deletar')){
+        return redirect()->route('admin.principal');
+      }
       User::find($id)->delete();
       \Session::flash('mensagem',[
         'msg'=> 'Registro deletado com sucesso',
@@ -88,4 +108,31 @@ class UsuarioController extends Controller
       ]);
       return redirect()->route('admin.usuarios');
     }
+
+
+    //Papel
+    public function papel($id)
+    {
+      $usuario = User::find($id);
+      $papel = Papel::all();
+      return view('admin.usuarios.papel', compact('usuario', 'papel'));
+    }
+
+    public function salvarPapel(Request $request, $id)
+    {
+      $usuario = User::find($id);
+      $dados = $request->all();
+      $papel = Papel::find($dados['papel_id']);
+      $usuario->adicionaPapel($papel);
+      return redirect()->back();
+    }
+
+    public function removerPapel($id, $papel_id)
+    {
+      $usuario = User::find($id);
+      $papel = Papel::find($papel_id);
+      $usuario->removerPapel($papel);
+      return redirect()->back();
+    }
+
 }
